pyproject.toml
[project]
name = "blog"
version = "1.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "beautifulsoup4>=4.13.4",
  "gitpython>=3.1.44",
  "pelican>=4.11.0",
  "pygments>=2.18.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
  "pytest>=8.4.1",
  "pytest-snapshot>=0.9.0",
  "pyyaml>=6.0.2",
  "ruff>=0.12.1",
]

[tool.pytest.ini_options]
testpaths = ["tests/tools"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
